Add metasprites sample and rand8/oam_meta_spr (#83)
Add a new samples/metasprites demo (Program.cs, metasprites.c, chr_generic.s, metasprites.csproj) that shows 16 bouncing 2x2 metasprites. Implement builtin subroutines in BuiltInSubroutines.cs: a Galois LFSR rand8 (and set_rand) using RAND_SEED, and an oam_meta_spr implementation to write metasprite data into the OAM buffer (uses TEMP/TEMP2 and PTR conventions). Includes minor supporting updates to task utilities and constants to integrate these new operations.

* Fix metasprites transpiler bugs and add tests

Fix three bugs discovered during testing:
- Stloc_s: always allocate addresses (not just when _runtimeValueInA),
  fixing 'index local has no address' error in HandleStelemI1
- Clear _runtimeValueInA after branch instructions (Bne_un_s, Brfalse_s)
  to prevent WriteLdc from skipping LDA when it shouldn't
- EmitOamMetaSpr: fix GetHashCode() != 0 check that would fail for
  local index 0 (use GetLdlocIndex directly instead)

Add metasprites test data (debug/release DLLs) and verified ROM output.

* Document metasprites sample in README

* Fix rand8 zero-seed bug and move RAND_SEED to $3C

- Add BEQ check in rand8 Galois LFSR to bootstrap from zero seed
  (ASL of 0 stays 0 forever; now skips to EOR #$CF when seed=0)
- Move RAND_SEED from $18 (between TEMP/TEMP2) to $3C (cc65 compatible)
  to avoid potential conflicts with TEMP usage in oam_meta_spr
- Update metasprites verified.bin for new rand8 output

* Fix pal_all/pal_bg byte array address loading and PR review fixes

- Fix critical bug: byte array address not loaded before pal_all/pal_bg
  when array stays on stack (not stored to a local). Add stack-marker
  detection in Call handler to emit LDA/LDX for unconsumed byte arrays.
- Guard AND RemoveLastInstructions against _runtimeValueInA=true
  (WriteLdc skips LDA when A has runtime value, nothing to remove)
- Fix set_rand inclusion: emit independently of rand8 usage
- Change set_rand API from ushort to byte (matches 1-byte RAND_SEED)
- Remove dead variable assignments (valueOps, indexIdx, arrayRef)
- Add metasprites project to samples/samples.sln
- Update ArrayIndexers test for correct byte array address loading
- Update verified.bin snapshots

* Enable warnings-as-errors; simplify IL lookahead

Enable TreatWarningsAsErrors in Directory.Build.props to treat compiler warnings as errors. In IL2NESWriter.cs, replace the forward loop that scanned instructions after ldc_i4_0/ldc_i4(_s) with a single immediate next-instruction check (i+1) and set andMask or subValue when the next opcode is And or Sub. This simplifies logic, avoids unnecessary iteration, and reduces risk of index-related issues.

* Fix local label scoping in Program6502 (oam_meta_spr JMP @loop bug)

Local labels (starting with @) were registered in a global label table,
causing collisions when multiple blocks used the same name (e.g., @loop
in both ppu_wait_frame and oam_meta_spr). The last block to register
would overwrite earlier addresses, making JMP @loop in oam_meta_spr
jump to the wrong subroutine's loop.

Fix: Scope local labels by prefixing with the block name during
registration and resolution. Add CurrentScope to LabelTable for
scoped lookups during emission.

* Fix CalculateFinalBuiltInsSize to match set_rand gating logic

* Remove reference ROM (should not be in repo)

+++ src/neslib/NESLib.cs
@@ -332,7 +332,7 @@ public static class NESLib
     /// <summary>
     /// set random seed
     /// </summary>
-    public static void set_rand(ushort seed) => throw null!;
+    public static void set_rand(byte seed) => throw null!;
 
     /// <summary>
     /// when display is enabled, vram access could only be done with this vram update system
